requirements.txt
# Core Data Science
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0

# Real-time Data & APIs
yfinance==0.2.18
alpha-vantage==2.3.0
requests==2.31.0
websocket-client==1.6.1

# Technical Analysis
ta-lib==0.4.26

# News & Sentiment
newspaper3k==0.2.8
textblob==0.17.1
vaderSentiment==3.3.2

# Backtesting & Trading
backtrader==1.9.78.123

# Visualization
plotly==5.15.0
dash==2.14.1
matplotlib==3.7.2

# Web Framework
flask==2.3.3
flask-socketio==5.3.6

# Utilities
python-dotenv==1.0.0
schedule==1.2.0